Rewritten get_material_ids to use the new file reading system
Moved Material class to misc
