+++ ACExplorer/ACUnity/__init__.py
@@ -10,7 +10,7 @@ from readFile_ import readFile
 from exportTexture_ import export_texture
 from exportModel_ import export_obj
 import formatFile
-from getMaterialIDs_ import getMaterialIDs
+from getMaterialIDs_ import get_material_ids
 
 from readModel_ import read_model
 from ACExplorer.ACUnity.exportFakes_ import export_fakes
