+++ ACExplorer/misc/__init__.py
@@ -1,4 +1,4 @@
-from texture import BaseTexture
+from texture import BaseTexture, Material
 from log import logger
 from decompress_ import decompress
 from tempFiles2 import tempFilesContainer
