+++ ACExplorer/misc/texture.py
@@ -47,3 +47,16 @@ class BaseTexture:
 		if self.imgDXT == 8:
 			texconv = '"{}" -fl 9.1 -y -px {}{} -f BC3_UNORM {}'.format(self.app.CONFIG['texconv'], self.app.CONFIG['dumpFolder'], os.sep, path)
 			os.system(texconv)
+
+
+class Material:
+	def __init__(self, name, missing_no=False, diffuse=None, normal=None, specular=None, height=None, transmission=None, mask1=None, mask2=None):
+		self.name = name
+		self.missing_no = missing_no
+		self.diffuse = diffuse
+		self.normal = normal
+		self.specular = specular
+		self.height = height
+		self.transmission = transmission
+		self.mask1 = mask1
+		self.mask2 = mask2
